Refactor namespaces and improve field handling logic
This commit refactors the codebase by changing the namespace from `Dimmer.DimmerSearch.AbstractQueryTree` and `Dimmer.DimmerSearch.AbstractQueryTree.NL` to `Dimmer.DimmerSearch.TQL` across multiple files.

Key changes include:
- Updated boolean handling in `AstEvaluator.cs` to include "oui" as a valid affirmative response.
- Adjusted field definitions in `FieldRegistry.cs`, moving "HasLyrics" and adding "PlainLyrics".
- Enhanced numeric value parsing for duration fields in `AstEvaluator.cs`.
- Cleaned up the overall structure by removing unnecessary using directives and improving namespace organization for better clarity and maintainability.
