+++ Dimmer/Dimmer/Dimmer/DimmerSearch/TQL/AstParser.cs
@@ -4,7 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Dimmer.DimmerSearch.AbstractQueryTree;
+namespace Dimmer.DimmerSearch.TQL;
 
 public class AstParser
 {
