+++ Dimmer/Dimmer/Dimmer/DimmerSearch/TQL/IQueryNode.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Dimmer.DimmerSearch.AbstractQueryTree;
+namespace Dimmer.DimmerSearch.TQL;
 public interface IQueryNode
 {
 }
