+++ Dimmer/Dimmer/Dimmer/DimmerSearch/TQL/LogicalNode.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Dimmer.DimmerSearch.AbstractQueryTree;
+namespace Dimmer.DimmerSearch.TQL;
 public class LogicalNode : IQueryNode
 {
     public IQueryNode Left { get; }
