+++ Dimmer/Dimmer/Dimmer/DimmerSearch/TQL/NaturalLanguageProcessor.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
-namespace Dimmer.DimmerSearch.AbstractQueryTree.NL;
+namespace Dimmer.DimmerSearch.TQL;
 
     public static class NaturalLanguageProcessor
     {
